New Changes

+++ MeetingRoom/src/com/interview/practice/domain/User.java
@@ -5,10 +5,10 @@ public class User {
     private final String userId;
     private final String name;
     private final String emailId;
-    private final int phoneNumber;
+    private final long phoneNumber;
 
 
-    public User(String userId,String name,String emailId,int phoneNumber){
+    public User(String userId,String name,String emailId,long phoneNumber){
         this.userId = userId;
         this.name = name;
         this.emailId = emailId;
